% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodegpole_phenology_event.R
\docType{data}
\name{lodgepole_phenology_event}
\alias{lodgepole_phenology_event}
\title{PINUS CONTORTA VAR LATIFOLIA FLOWERING PHENOLOGY EVENTS}
\format{
A data frame with 15707 rows and 15 variables:
\describe{
  \item{\code{Index}}{integer An index uniquely identifying a tree from a clone observed in a year at an orchard on a site}
  \item{\code{State}}{integer Recorded phenophases translated into 1 - not yet flowering, 2 - flowering, or 3 - done flowering. }
  \item{\code{Year}}{integer Year of observation}
  \item{\code{Event_Obs}}{integer Observations tied to events 1 - before flowering (the last day a tree was observed in state 1), 2 - first flowering (the first day a tree was observed in state 2), 3 - last flowering (the last day a tree was observed in state 2), and 4 - past flowering (the first day a tree was observed in state 3)}
  \item{\code{DoY}}{integer Day of year of observation}
  \item{\code{Event_Label}}{character Text descriptions of events/observation types}
  \item{\code{Date}}{character Date tree observed}
  \item{\code{Sex}}{character MALE or FEMALE strobili}
  \item{\code{Site}}{character Seed orchard location name}
  \item{\code{Orchard}}{integer Number referencing an orchard at a specific seed orchard location (site). Corresponds to the provenance of the trees grown in that orchard}
  \item{\code{Clone}}{integer Clone identifier}
  \item{\code{Tree}}{integer Tree identifier - not all trees have identifiers and many are NA}
  \item{\code{X}}{character X location of tree in orchard}
  \item{\code{Y}}{integer Y location of tree in orchard}
  \item{\code{Source}}{character Data provided by Rita Wagner working at PGTIS or Chris Walsh working at Kalamalka}
}
}
\usage{
lodgepole_phenology_event
}
\description{
Phenology of pollen shed and cone receptivity ("flowering") in lodgepole pine. Years cover 1997-2012. Data was collected at seven sites with multiple provenances represented at many of those sites. All sites were British Columbia seed orchard locations. There are multiple orchards at each site. Each orchard has trees whose seeds were sourced from a specific seed planning unit (as defined by the ministry in British Columbia.)
This data set presents the data as observations of phenological states (see lodgepole_phenology_state) translated into proxies for phenological events: 1 - before flowering (the last day a tree was observed in state 1), 2 - first flowering (the first day a tree was observed in state 2), 3 - last flowering (the last day a tree was observed in state 2), and 4 - past flowering (the first day a tree was observed in state 3). Trees that did not produce strobili are *excluded* from this dataset
}
\keyword{datasets}
