% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodgepole_phenology_state.R
\docType{data}
\name{lodgepole_phenology_state}
\alias{lodgepole_phenology_state}
\title{PINUS CONTORTA VAR LATIFOLIA FLOWERING PHENOLOGY STATES}
\format{
A data frame with 29058 rows and 14 variables:
\describe{
  \item{\code{Index}}{integer An index uniquely identifying a tree from a clone observed in a year at an orchard on a site}
  \item{\code{DoY}}{integer Day of year of observation}
  \item{\code{Date}}{character Date state was observed}
  \item{\code{Phenophase_Recorded}}{character Phenological state as recorded by workers in the field}
  \item{\code{State}}{integer integer Recorded phenophases translated into 1 - not yet flowering, 2 - flowering, or 3 - done flowering. 0 means the tree did not produce strobli that year}
  \item{\code{Sex}}{character MALE or FEMALE strobili}
  \item{\code{Year}}{integer Year of observation}
  \item{\code{Site}}{character Seed orchard location name}
  \item{\code{Orchard}}{integer Number referencing an orchard at a specific seed orchard location (site). Corresponds to the provenance of the trees grown in that orchard}
  \item{\code{Clone}}{integer Clone identifier}
  \item{\code{Tree}}{integer Tree identifier - not all trees have identifiers and many are NA}
  \item{\code{X}}{character X location of tree in orchard}
  \item{\code{Y}}{integer Y location of tree in orchard}
  \item{\code{Source}}{character Data provided by Rita Wagner working at PGTIS or Chris Walsh working at Kalamalka}
}
}
\usage{
lodgepole_phenology_state
}
\description{
Phenology of pollen shed and cone receptivity ("flowering") in lodgepole pine. Years cover 1997-2012. Data was collected at seven sites with multiple provenances represented at many of those sites. All sites were British Columbia seed orchard locations. There are multiple orchards at each site. Each orchard has trees whose seeds were sourced from a specific seed planning unit (as defined by the ministry in British Columbia.)
This data set presents the data as observations of phenological states (0 - tree did not produce flowers, 1 - tree is before flowering, 2 - tree is flowering, 3 - tree is past flowering)
}
\keyword{datasets}
