% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\docType{data}
\name{phenology}
\alias{phenology}
\title{PINUS CONTORTA VAR LATIFOLIA FLOWERING PHENOLOGY}
\format{
A data frame with 36182 rows and 16 variables:
\describe{
  \item{\code{Index}}{integer COLUMN_DESCRIPTION}
  \item{\code{DoY}}{integer Day of year}
  \item{\code{Phenophase}}{character Phenophase as recorded by workers in the field}
  \item{\code{Sex}}{character Male or Female strobili}
  \item{\code{Year}}{integer Year}
  \item{\code{Site}}{character Seed orchard location name}
  \item{\code{Orchard}}{integer Number referencing an orchard at a specific seed orchard location (site). Corresponds to the provenance of the trees grown in that orchard}
  \item{\code{Clone}}{integer Clone identifier}
  \item{\code{Tree}}{integer Tree identifier - not all trees have identifiers}
  \item{\code{X}}{character X location of tree in orchard}
  \item{\code{Y}}{integer Y location of tree in orchard}
  \item{\code{Date}}{character Date phenology observed}
  \item{\code{Source}}{character Data provided by Rita Wagner or Chris Walsh}
  \item{\code{First_RF}}{integer Day strobili first recorded flowering}
  \item{\code{Last_RF}}{integer Day strobili last recorded flowering}
  \item{\code{Phenophase_Derived}}{integer Recorded phenophases translated into 1 - not yet flowering, 2 - flowering, or 3 - done flowering}
}
}
\usage{
phenology
}
\description{
Phenology of pollen shed and cone receptivity in lodgepole pine. Years cover 1997-2011. Data was collected at seven sites with multiple provenances represented at (most) sites. All sites were British Columbia seed orchard locations. There are multiple orchards at each site. Each orchard has trees whose seeds were sourced from a specific seed planning unit (as defined by the ministry in British Columbia.)
}
\keyword{datasets}
