% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotflower.R
\name{plot_flower}
\alias{plot_flower}
\title{Flower plot}
\usage{
plot_flower(.Data, title = NA, legend_include = TRUE, filename = NA)
}
\arguments{
\item{.Data}{data frame containing scores to be plotted. Column names should be
"pos", "score", "weight", "goal", "name_supra", "pos_supra"}

\item{title}{optional title for the plot}

\item{legend_include}{logical, whether to include a plot legend, defaults to TRUE}

\item{filename}{if not NA, save the figure using this filename (relative or absolute)}
}
\value{
ggplot object of the flowerplot
}
\description{
Flower plot
}
\examples{

}
