% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotflower.R
\name{plot_flower}
\alias{plot_flower}
\title{Flower plot}
\usage{
plot_flower(.Data, title = NA, legend_include = TRUE, colors = NA,
  fixed_colors = FALSE, filename = NA)
}
\arguments{
\item{.Data}{data frame containing scores to be plotted. Column names should include
"score", "weight", "name_supra", and "name_flower"}

\item{title}{optional title for the plot}

\item{legend_include}{logical, whether to include a plot legend, defaults to TRUE}

\item{colors}{an optional color palette to be used for the petal colors}

\item{fixed_colors}{if TRUE, then use a discrete fixed color palette for coloring petals
based on petal categories; defaults to FALSE,}

\item{filename}{if not NA, save the figure using this filename (relative or absolute)}
}
\value{
ggplot object of the flowerplot
}
\description{
Flower plot
}
\examples{

data(ohi)
plot_flower(ohi, "OHI Example")

}
